build/jenkins: refactor archives
Split out the binary archive handling into a separate file. Their
behaviour must be the same on all backends. Putting them into a single
file makes their extension and maintenance considerably easier.
